Refactored ActionType
