+++ src/main/java/inf112/skeleton/app/GUI/BaseAsset.java
@@ -1,4 +1,5 @@
-package com.jarlescene2d;
+package inf112.skeleton.app.GUI;
+
 
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.graphics.Texture;
