+++ src/main/java/inf112/skeleton/app/GUI/pieces/Robot.java
@@ -2,6 +2,7 @@ package inf112.skeleton.app.GUI.pieces;
 
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.graphics.Texture;
+import com.badlogic.gdx.graphics.g2d.Batch;
 import com.badlogic.gdx.graphics.g2d.Sprite;
 
 
@@ -22,9 +23,10 @@ public class Robot extends Piece {
                 sprite = new Sprite(new Texture(Gdx.files.internal(path)));
         }
 
-        /*@Override
+
+        @Override
         public void draw(Batch batch, float parentAlpha) {
            //batch.enableBlending();
             batch.draw(sprite, getX(),getY(), getWidth(), getHeight());
-        }*/
+        }
 }
