+++ src/main/java/inf112/skeleton/app/gameLogic/Action.java
@@ -1,6 +1,6 @@
 package inf112.skeleton.app.gameLogic;
 
-import inf112.skeleton.app.Game.Enum.ActionType;
+import inf112.skeleton.app.gameLogic.enums.ActionType;
 import inf112.skeleton.app.gameLogic.Player;
 import inf112.skeleton.app.gameLogic.ProgramCard;
 import inf112.skeleton.app.gameLogic.enums.Rotation;
