+++ src/main/java/inf112/skeleton/app/gameLogic/board/activeCells/Gears.java
@@ -1,6 +1,6 @@
 package inf112.skeleton.app.gameLogic.board.activeCells;
 
-import inf112.skeleton.app.Game.Enum.ActionType;
+import inf112.skeleton.app.gameLogic.enums.ActionType;
 import inf112.skeleton.app.gameLogic.enums.Rotation ;
 import inf112.skeleton.app.gameLogic.board.IActiveCell ;
