+++ src/main/java/inf112/skeleton/app/gameLogic/enums/ActionType.java
@@ -1,4 +1,4 @@
-package inf112.skeleton.app.Game.Enum;
+package inf112.skeleton.app.gameLogic.enums;
 
 public enum ActionType {
     DAMAGE, MOVE, TURN
