+++ src/main/java/inf112/skeleton/app/gameLogic/game/Game.java
@@ -11,6 +11,6 @@ public class Game {
     }
 
     public void doPhase1(){
-        
+
     }
 }
